+++ Machine/include/Machine/IMachine.h
@@ -23,6 +23,7 @@ SOFTWARE.
 #ifndef IMACHINE_H
 #define IMACHINE_H
 
+#include <functional>
 #include <memory>
 #include <string>
 #include "Controller/IController.h"
@@ -158,6 +159,54 @@ namespace MachEmu
 		virtual ErrorCode SetOptions(const char* options) = 0;
 
 		/** Save the state of the machine.
+		
+			Upon saving the state of the machine, call the method `onSave` to process the save state.
+		
+			The json layout of an example save state is specifed in the json snippet below:
+
+			@code{.json}
+
+			{
+				"cpu":
+				{
+					"uuid":"O+hPH516S3ClRdnzSRL8rQ==",	// The base64 unique identifier of the cpu
+					"registers":
+					{
+						"a":0,							// The a register
+						"b":0,							// The b register
+						"c":0,							// The c register
+						"d":0,							// The d register
+						"e":0,							// The e register
+						"h":0,							// The h register
+						"l":0,							// The l register
+						"s":2							// The status register
+					},
+					"pc":0,								// The program counter
+					"sp":0								// The stack pointer
+				},
+			{
+			"memory":
+			{
+				"uuid":"zRjYZ92/TaqtWroc666wMQ==",		// The base64 unique identifier for the memory controller
+				"rom":"FkgjfhUYrudiMj7y65789Io=",		// The base64 MD5 hash of the rom
+				"ram":
+				{
+					"encoder":"base64",					// The binary to text encoding for the ram
+					"compressor":"zlib",				// The compressor to use for the ram
+					"size":57343,						// The size of the uncompressed ram
+					"bytes":"... the ram bytes ..."		// The compressed and encoded ram bytes
+				}
+			}
+
+			@endcode
+
+			@param	onSave		The method to call once the save state has been processed.
+
+			@since version 1.5.0
+		*/
+		virtual void OnSave(std::function<void(std::string&& json)>&& onSave) = 0;
+
+		/** Save the state of the machine.
 
 			Returns the state of the machine as a JSON string.
 
@@ -181,8 +230,10 @@ namespace MachEmu
 			<tr><td>pc</td><td>Contents of the program counter</td></tr>
 			<tr><td>sp</td><td>Contents of the stack pointer</td></tr>
 			</table>
+
+			@deprecated	since 1.5.0
 		*/
-		virtual std::string Save() const = 0;
+		[[deprecated("Will be removed in v2.0.0, please use OnSave")]] virtual std::string Save() const = 0;
 
 		/**	Set the frequency at which the internal clock ticks.