+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 * Added `encoder` config option.
 * Added `compressor` config option.
 * Added cmake option `enableZlib`.
+* Deprecated `IMachine::Save`.
+* Added Machine interface method `OnSave`.
+* Added `ramOffset` type config option.
+* Added `ramSize` type config option.
+* Added `romOffset` type config option.
+* Added `romSize` type config option.
 
 1.4.0 [19/03/24]
 * Added Version API method.