+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-1.5.0
+1.5.0 [19/04/24]
 * Added Machine interface method `OnLoad`.
 * Added Machine interface method `OnSave`.
 * Added Controller interface method `Uuid`.